♻️  Reorganisation des specs GF (#2021)
* ♻️ Reorganize garanties financières specs

* 🎨 Retours reviews
Update packages/specifications/src/projet/lauréat/garantiesFinancières/actuelles/enregistrerGarantiesFinancièresActuelles.feature


Update packages/specifications/src/projet/lauréat/garantiesFinancières/actuelles/importerTypeGarantiesFinancièresActuelles.feature


Update packages/specifications/src/projet/lauréat/garantiesFinancières/actuelles/modifierGarantiesFinancièresActuelles.feature


Update packages/specifications/src/projet/lauréat/garantiesFinancières/actuelles/enregistrerAttestationGarantiesFinancièresActuelles.feature


Update packages/specifications/src/projet/lauréat/garantiesFinancières/actuelles/enregistrerGarantiesFinancièresActuelles.feature


Update packages/specifications/src/projet/lauréat/garantiesFinancières/dépôt/soumettreDépôtGarantiesFinancières.feature


Update packages/specifications/src/projet/lauréat/garantiesFinancières/dépôt/soumettreDépôtGarantiesFinancières.feature


Update packages/specifications/src/projet/lauréat/garantiesFinancières/dépôt/soumettreDépôtGarantiesFinancières.feature


🐛 Fix tests

* 🔥 Suppression du step "la liste des dépôts de garanties financières devrait être vide" puisqu'on va tester ça dans le  scénario "il ne devrait pas y avoir de dépôt de garanties financières pour le projet"
