+++ packages/specifications/src/notification/stepDefinitions/notification.then.ts
@@ -6,9 +6,9 @@ import { PotentielWorld } from '../../potentiel.world';
 Alors(
   'un email a été envoyé à {string} avec :',
   function (this: PotentielWorld, email: string, data: DataTable) {
-    const example = data.rowsHash();
-    const notif = this.notificationWorld.récupérerNotification(email, example.sujet);
-    for (const [key, value] of Object.entries(example)) {
+    const exemple = data.rowsHash();
+    const notif = this.notificationWorld.récupérerNotification(email, exemple.sujet);
+    for (const [key, value] of Object.entries(exemple)) {
       if (key === 'sujet') {
         expect(notif.messageSubject).to.equal(value);
       } else {
