+++ packages/specifications/src/projet/lauréat/garantiesFinancières/effacerHistorique/stepDefinitions/effacerHistoriqueGarantiesFinancières.when.ts
@@ -1,0 +1,28 @@
+import { When as Quand } from '@cucumber/cucumber';
+import { mediator } from 'mediateur';
+
+import { GarantiesFinancières } from '@potentiel-domain/laureat';
+
+import { PotentielWorld } from '../../../../../potentiel.world';
+import { sleep } from '../../../../../helpers/sleep';
+
+Quand(
+  `un admin efface l'historique des garanties financières pour le projet {string}`,
+  async function (this: PotentielWorld, nomProjet: string) {
+    try {
+      const { identifiantProjet } = this.lauréatWorld.rechercherLauréatFixture(nomProjet);
+
+      await mediator.send<GarantiesFinancières.EffacerHistoriqueGarantiesFinancièresUseCase>({
+        type: 'Lauréat.GarantiesFinancières.UseCase.EffacerHistoriqueGarantiesFinancières',
+        data: {
+          identifiantProjetValue: identifiantProjet.formatter(),
+          effacéLeValue: new Date().toISOString(),
+          effacéParValue: 'admin@test.test',
+        },
+      });
+      await sleep(500);
+    } catch (error) {
+      this.error = error as Error;
+    }
+  },
+);
