+++ packages/specifications/src/tâche-planifiée/annulerTâchePlanifiée.feature
@@ -9,7 +9,7 @@ Fonctionnalité: Annuler une tâche planifiée
             | role  | porteur-projet      |
 
     Scénario: Une tâche du type "échoir les garanties financières" est annulée quand une attestation de conformité est transmise
-        Etant donné des garanties financières validées pour le projet "Du boulodrome de Marseille" avec :
+        Etant donné des garanties financières actuelles pour le projet "Du boulodrome de Marseille" avec :
             | type               | avec-date-échéance |
             | date d'échéance    | 2024-12-01         |
             | date de validation | 2024-11-24         |
@@ -18,7 +18,7 @@ Fonctionnalité: Annuler une tâche planifiée
         Alors une tâche "échoir les garanties financières" n'est plus planifiée pour le projet "Du boulodrome de Marseille"
 
     Scénario: Une tâche du type "rappel échéance garanties financières à un mois" est annulée quand une attestation de conformité est transmise
-        Etant donné des garanties financières validées pour le projet "Du boulodrome de Marseille" avec :
+        Etant donné des garanties financières actuelles pour le projet "Du boulodrome de Marseille" avec :
             | type               | avec-date-échéance |
             | date d'échéance    | 2024-12-01         |
             | date de validation | 2024-11-24         |
@@ -27,7 +27,7 @@ Fonctionnalité: Annuler une tâche planifiée
         Alors une tâche "rappel échéance garanties financières à un mois" n'est plus planifiée pour le projet "Du boulodrome de Marseille"
 
     Scénario: Une tâche du type "rappel échéance garanties financières à deux mois" est annulée quand une attestation de conformité est transmise
-        Etant donné des garanties financières validées pour le projet "Du boulodrome de Marseille" avec :
+        Etant donné des garanties financières actuelles pour le projet "Du boulodrome de Marseille" avec :
             | type               | avec-date-échéance |
             | date d'échéance    | 2024-12-01         |
             | date de validation | 2024-11-24         |
@@ -36,28 +36,28 @@ Fonctionnalité: Annuler une tâche planifiée
         Alors une tâche "rappel échéance garanties financières à deux mois" n'est plus planifiée pour le projet "Du boulodrome de Marseille"
 
     Scénario: Une tâche du type "échoir les garanties financières" est annulée quand une dépôt de garanties financiéres est soumis
-        Etant donné des garanties financières validées pour le projet "Du boulodrome de Marseille" avec :
+        Etant donné des garanties financières actuelles pour le projet "Du boulodrome de Marseille" avec :
             | type            | avec-date-échéance |
             | date d'échéance | 2024-12-01         |
-        Quand un porteur soumet des garanties financières pour le projet "Du boulodrome de Marseille" avec :
+        Quand un porteur soumet un dépôt de garanties financières pour le projet "Du boulodrome de Marseille" avec :
             | type            | avec-date-échéance |
             | date d'échéance | 2024-12-01         |
         Alors une tâche "échoir les garanties financières" n'est plus planifiée pour le projet "Du boulodrome de Marseille"
 
     Scénario: Une tâche du type "rappel échéance garanties financières à un mois" est annulée quand une attestation de conformité est transmise
-        Etant donné des garanties financières validées pour le projet "Du boulodrome de Marseille" avec :
+        Etant donné des garanties financières actuelles pour le projet "Du boulodrome de Marseille" avec :
             | type            | avec-date-échéance |
             | date d'échéance | 2024-12-01         |
-        Quand un porteur soumet des garanties financières pour le projet "Du boulodrome de Marseille" avec :
+        Quand un porteur soumet un dépôt de garanties financières pour le projet "Du boulodrome de Marseille" avec :
             | type            | avec-date-échéance |
             | date d'échéance | 2024-12-01         |
         Alors une tâche "rappel échéance garanties financières à un mois" n'est plus planifiée pour le projet "Du boulodrome de Marseille"
 
     Scénario: Une tâche du type "rappel échéance garanties financières à deux mois" est annulée quand une attestation de conformité est transmise
-        Etant donné des garanties financières validées pour le projet "Du boulodrome de Marseille" avec :
+        Etant donné des garanties financières actuelles pour le projet "Du boulodrome de Marseille" avec :
             | type            | avec-date-échéance |
             | date d'échéance | 2024-12-01         |
-        Quand un porteur soumet des garanties financières pour le projet "Du boulodrome de Marseille" avec :
+        Quand un porteur soumet un dépôt de garanties financières pour le projet "Du boulodrome de Marseille" avec :
             | type            | avec-date-échéance |
             | date d'échéance | 2024-12-01         |
         Alors une tâche "rappel échéance garanties financières à deux mois" n'est plus planifiée pour le projet "Du boulodrome de Marseille"
